Renaming utils/highlight -> utils/cterm_colors
