+++ lua/lualine/highlight.lua
@@ -2,13 +2,13 @@
 -- MIT license, see LICENSE for more details.
 
 local M = {  }
-local utils_hl = require "lualine.utils.highlight"
+local utils_colors = require "lualine.utils.cterm_colors"
 
 local function highlight (name, foreground, background, gui)
   local command = {
       'highlight', name,
-      'ctermfg=' .. (foreground[2] or utils_hl.get_cterm_color(foreground)),
-      'ctermbg=' .. (background[2] or utils_hl.get_cterm_color(background)),
+      'ctermfg=' .. (foreground[2] or utils_colors.get_cterm_color(foreground)),
+      'ctermbg=' .. (background[2] or utils_colors.get_cterm_color(background)),
       'cterm=' .. (gui or 'none'),
       'guifg=' .. (foreground[1] or foreground),
       'guibg=' .. (background[1] or background),
